+++ src/unfold/mixins/formfield_model_admin.py
@@ -1,12 +1,16 @@
 import copy
 from typing import Any
 
-from django.contrib.admin import helpers
+from django.contrib.admin.options import BaseModelAdmin
 from django.contrib.admin.sites import AdminSite
-from django.contrib.admin.widgets import RelatedFieldWidgetWrapper
+from django.contrib.admin.widgets import (
+    FilteredSelectMultiple,
+    RelatedFieldWidgetWrapper,
+)
 from django.db import models
 from django.db.models.fields import Field
 from django.db.models.fields.related import ForeignKey, ManyToManyField
+from django.forms.fields import Field as FormField
 from django.forms.fields import TypedChoiceField
 from django.forms.models import ModelChoiceField, ModelMultipleChoiceField
 from django.forms.widgets import SelectMultiple
@@ -17,7 +21,11 @@ from unfold import widgets
 from unfold.overrides import FORMFIELD_OVERRIDES
 
 
-class BaseModelAdminMixin:
+class FormFieldModelAdminMixin(BaseModelAdmin):
+    # List of all db fields which are not available in autocomplete_fields
+    _autocomplete_fields_missing: list[str] = []
+    autocomplete_fields_excluded_from_warnings: list[str] = []
+
     def __init__(self, model: type[models.Model], admin_site: AdminSite) -> None:
         overrides = copy.deepcopy(FORMFIELD_OVERRIDES)
 
@@ -28,22 +36,9 @@ class BaseModelAdminMixin:
 
         super().__init__(model, admin_site)
 
-    def changeform_view(
-        self,
-        request: HttpRequest,
-        object_id: str | None = None,
-        form_url: str = "",
-        extra_context: dict[str, Any] | None = None,
-    ) -> Any:
-        from unfold.forms import AdminForm, Fieldline
-
-        helpers.AdminForm = AdminForm
-        helpers.Fieldline = Fieldline
-        return super().changeform_view(request, object_id, form_url, extra_context)
-
     def formfield_for_choice_field(
-        self, db_field: Field, request: HttpRequest, **kwargs
-    ) -> TypedChoiceField:
+        self, db_field: Field, request: HttpRequest, **kwargs: Any
+    ) -> TypedChoiceField | None:
         if "widget" not in kwargs:
             if db_field.name in self.radio_fields:
                 kwargs["widget"] = widgets.UnfoldAdminRadioSelectWidget(
@@ -60,15 +55,13 @@ class BaseModelAdminMixin:
         return super().formfield_for_choice_field(db_field, request, **kwargs)
 
     def formfield_for_foreignkey(
-        self, db_field: ForeignKey, request: HttpRequest, **kwargs
+        self, db_field: ForeignKey, request: HttpRequest, **kwargs: Any
     ) -> ModelChoiceField | None:
-        db = kwargs.get("using")
-
         # Overrides widgets for all related fields
         if "widget" not in kwargs:
             if db_field.name in self.raw_id_fields:
                 kwargs["widget"] = widgets.UnfoldForeignKeyRawIdWidget(
-                    db_field.remote_field, self.admin_site, using=db
+                    db_field.remote_field, self.admin_site, using=kwargs.get("using")
                 )
             elif (
                 db_field.name not in self.get_autocomplete_fields(request)
@@ -77,36 +70,35 @@ class BaseModelAdminMixin:
                 kwargs["widget"] = widgets.UnfoldAdminSelectWidget()
                 kwargs["empty_label"] = _("Select value")
 
-        return super().formfield_for_foreignkey(db_field, request, **kwargs)
+        formfield = super().formfield_for_foreignkey(db_field, request, **kwargs)
+        self._check_autocomplete_field(db_field, formfield, request)
+        return formfield
 
     def formfield_for_manytomany(
         self,
         db_field: ManyToManyField,
         request: HttpRequest,
-        **kwargs,
-    ) -> ModelMultipleChoiceField:
-        if "widget" not in kwargs:
-            if db_field.name in self.raw_id_fields:
-                kwargs["widget"] = widgets.UnfoldAdminTextInputWidget()
-
-        form_field = super().formfield_for_manytomany(db_field, request, **kwargs)
+        **kwargs: Any,
+    ) -> ModelMultipleChoiceField | None:
+        formfield = super().formfield_for_manytomany(db_field, request, **kwargs)
 
         # If M2M uses intermediary model, form_field will be None
-        if not form_field:
+        if not formfield:
             return None
 
-        if isinstance(form_field.widget, SelectMultiple):
-            form_field.widget.attrs["class"] = " ".join(widgets.SELECT_CLASSES)
+        if isinstance(formfield.widget, SelectMultiple):
+            formfield.widget.attrs["class"] = " ".join(widgets.SELECT_CLASSES)
 
-        return form_field
+        self._check_autocomplete_field(db_field, formfield, request)
+        return formfield
 
     def formfield_for_nullboolean_field(
-        self, db_field: Field, request: HttpRequest, **kwargs
-    ) -> Field | None:
+        self, db_field: Field, request: HttpRequest, **kwargs: Any
+    ) -> FormField | None:
         if "widget" not in kwargs:
             if db_field.choices:
                 kwargs["widget"] = widgets.UnfoldAdminSelectWidget(
-                    choices=db_field.choices
+                    choices=list(db_field.choices)
                 )
             else:
                 kwargs["widget"] = widgets.UnfoldAdminNullBooleanSelectWidget()
@@ -114,8 +106,8 @@ class BaseModelAdminMixin:
         return db_field.formfield(**kwargs)
 
     def formfield_for_dbfield(
-        self, db_field: Field, request: HttpRequest, **kwargs
-    ) -> Field | None:
+        self, db_field: Field, request: HttpRequest, **kwargs: Any
+    ) -> FormField | None:
         if isinstance(db_field, models.BooleanField) and db_field.null is True:
             return self.formfield_for_nullboolean_field(db_field, request, **kwargs)
 
@@ -127,3 +119,35 @@ class BaseModelAdminMixin:
             )
 
         return formfield
+
+    def _check_autocomplete_field(  # noqa: PLR0911
+        self,
+        db_field: Field,
+        formfield: ModelChoiceField | ModelMultipleChoiceField | None,
+        request: HttpRequest,
+    ) -> None:
+        # Field is already in autocomplete_fields
+        if db_field.name in self.get_autocomplete_fields(request):
+            return
+
+        # Readonly fields are not rendering large select dropdown
+        if db_field.name in self.get_readonly_fields(request):
+            return
+
+        # Raw ID field, no problem with SQL queries
+        if db_field.name in self.raw_id_fields:
+            return
+
+        # Make an exception for this special widget
+        if formfield is not None and isinstance(
+            formfield.widget, FilteredSelectMultiple
+        ):
+            return
+
+        # Sometimes we want to exclude a field from the warnings
+        if db_field.name in self.autocomplete_fields_excluded_from_warnings:
+            return
+
+        self._autocomplete_fields_missing.append(
+            f"{self.__class__.__name__}.{db_field.name}"
+        )