+++ src/unfold/settings.py
@@ -7,6 +7,7 @@ from unfold.widgets import (
     CHECKBOX_CLASSES,
     FILE_CLASSES,
     INPUT_CLASSES,
+    PROSE_CLASSES,
     RADIO_CLASSES,
     SWITCH_CLASSES,
 )
@@ -24,10 +25,12 @@ CONFIG_DEFAULTS = {
     "SHOW_HISTORY": True,
     "SHOW_VIEW_ON_SITE": True,
     "SHOW_LANGUAGES": False,
-    "LANGUAGE_FLAGS": {},
     "SHOW_BACK_BUTTON": False,
+    "SHOW_UI_WARNINGS": False,
+    "LANGUAGE_FLAGS": {},
     "FORMS": {
         "classes": {
+            "prose": " ".join(PROSE_CLASSES),
             "text_input": " ".join(INPUT_CLASSES),
             "checkbox": " ".join(CHECKBOX_CLASSES),
             "button": " ".join(BUTTON_CLASSES),