+++ src/unfold/static/unfold/js/app.js
@@ -21,7 +21,35 @@ window.addEventListener("load", (e) => {
  *************************************************************/
 function theme(defaultTheme = "auto") {
   return {
+    openModal: false,
+    filterOpen: false,
+    openAllApplications: false,
     adminTheme: Alpine.$persist(defaultTheme).as('adminTheme'),
+    init() {
+        this.$watch('openModal', (value) => {
+          if (value) {
+            document.getElementsByTagName("body")[0].classList.add("overflow-hidden");
+          } else {
+            document.getElementsByTagName("body")[0].classList.remove("overflow-hidden");
+          }
+        });
+
+        this.$watch('filterOpen', (value) => {
+          if (value) {
+            document.getElementsByTagName("body")[0].classList.add("overflow-hidden");
+          } else {
+            document.getElementsByTagName("body")[0].classList.remove("overflow-hidden");
+          }
+        });
+
+        this.$watch('openAllApplications', (value) => {
+          if (value) {
+            document.getElementsByTagName("body")[0].classList.add("overflow-hidden");
+          } else {
+            document.getElementsByTagName("body")[0].classList.remove("overflow-hidden");
+          }
+        });
+    },
     switchTheme(theme) {
       this.adminTheme = theme;
     },
@@ -351,7 +379,7 @@ function searchCommand() {
         this.scrollToActiveItem();
       }
     },
-    selectItem(addHistory) {
+    selectItem(addHistory, openInNewTab=false) {
       const link = this.items[this.currentIndex - 1].querySelector("a");
       const data = {
         title: link.dataset.title,
@@ -364,7 +392,11 @@ function searchCommand() {
         this.addToHistory(data);
       }
 
-      window.location = link.href;
+      if (openInNewTab) {
+        window.open(link.href, "_blank");
+      } else {
+        window.location = link.href;
+      }
     },
     addToHistory(data) {
       let commandHistory = JSON.parse(