+++ src/unfold/templatetags/unfold_list.py
@@ -2,8 +2,9 @@ import datetime
 from collections.abc import Generator
 from typing import Any
 
+from django import VERSION as DJANGO_VERSION
 from django.contrib.admin import SimpleListFilter
-from django.contrib.admin.options import IS_POPUP_VAR
+from django.contrib.admin.options import IS_FACETS_VAR, IS_POPUP_VAR
 from django.contrib.admin.templatetags.admin_list import (
     ResultList,
     _coerce_field_name,  # ty:ignore[unresolved-import]
@@ -36,17 +37,10 @@ from unfold.utils import (
 from unfold.views import DatasetChangeList
 from unfold.widgets import UnfoldBooleanWidget
 
-try:
-    from django.contrib.admin.options import IS_FACETS_VAR
-except ImportError:
-    # TODO: remove once django 4.x is not supported
-    IS_FACETS_VAR: str | None = None
-
 register = Library()
 
 LINK_CLASSES = [
-    "text-font-important-light",
-    "dark:text-font-important-dark",
+    "text-important",
 ]
 
 ROW_CLASSES = [
@@ -407,6 +401,14 @@ def result_list(context: dict[str, Any], cl: ChangeList) -> dict[str, Any]:
 
 @register.tag(name="unfold_result_list")
 def result_list_tag(parser: Parser, token: Token) -> InclusionAdminNode:
+    if DJANGO_VERSION >= (6, 1):
+        return InclusionAdminNode(
+            "unfold_result_list",
+            parser,
+            token,
+            func=result_list,
+            template_name="change_list_results.html",
+        )
     return InclusionAdminNode(
         parser,
         token,
@@ -526,6 +528,15 @@ def unfold_vertical_filters(cl: ChangeList) -> list[SimpleListFilter]:
 
 @register.tag(name="unfold_search_form")
 def unfold_search_form_tag(parser, token):
+    if DJANGO_VERSION >= (6, 1):
+        return InclusionAdminNode(
+            "unfold_search_form",
+            parser,
+            token,
+            func=unfold_search_form,
+            template_name="search_form.html",
+            takes_context=False,
+        )
     return InclusionAdminNode(
         parser,
         token,
@@ -537,6 +548,14 @@ def unfold_search_form_tag(parser, token):
 
 @register.tag(name="unfold_admin_actions")
 def unfold_admin_actions_tag(parser, token):
+    if DJANGO_VERSION >= (6, 1):
+        return InclusionAdminNode(
+            "unfold_admin_actions",
+            parser,
+            token,
+            func=admin_actions,
+            template_name="dataset_actions.html",
+        )
     return InclusionAdminNode(
         parser,
         token,